autochecks: bundle migration code III
